requirements.txt
annotated-types==0.7.0
anyio==4.10.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
bcrypt==4.3.0
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.110.0
fastapi-users==14.0.1
graphql-core==3.2.6
greenlet==3.2.4
h11==0.16.0
idna==3.10
iniconfig==2.1.0
makefun==1.16.0
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.10.1
pytest==8.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.36.3
strawberry-graphql==0.214.0
typing_extensions==4.14.1
uvicorn==0.29.0